+++ root-project/CHANGELOG.md
@@ -2,6 +2,23 @@
 
 ## Unreleased
 
+## `0.3.4` - 2020-03-25
+
+### Changed
+
+- Upgraded to GDK for Unity version `0.3.4`
+### Breaking Changes
+
+- The blank project now uses [query-based interest](https://docs.improbable.io/reference/14.4/shared/authority-and-interest/interest/query-based-interest-qbi) instead of [chunk-based interest](https://docs.improbable.io/reference/14.4/shared/authority-and-interest/interest/chunk-based-interest-cbi). [#89](https://github.com/spatialos/gdk-for-unity-blank-project/pull/89)
+- You must now pass in an `EntityId` to create a player `EntityTemplate`. [#90](https://github.com/spatialos/gdk-for-unity-blank-project/pull/90)
+
+## `0.3.3` - 2020-02-14
+
+### Changed
+
+- Upgraded to GDK for Unity version `0.3.3`
+- Updated the project app identifiers to `io.improbable.gdk.blankproject`. [#81](https://github.com/spatialos/gdk-for-unity-blank-project/pull/81)
+
 ## `0.3.2` - 2019-12-23
 
 ### Changed